+++ src/main/java/com/example/alba_pocket/repository/LikesRepository.java
@@ -14,6 +14,6 @@ public interface LikesRepository extends JpaRepository<Likes, Long> {
 
     boolean existsByUserIdAndPostId(Long userId, Long postId);
 
-    int countByPostId(Long postId);
+    Long countByPostId(Long postId);
 
 }
